Move utility function (#552)
