+++ llms/googleai/googleai.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/google/generative-ai-go/genai"
+	"github.com/tmc/langchaingo/internal/util"
 	"github.com/tmc/langchaingo/llms"
 	"github.com/tmc/langchaingo/schema"
 	"google.golang.org/api/iterator"
@@ -127,7 +128,7 @@ func convertParts(parts []llms.ContentPart) ([]genai.Part, error) {
 		case llms.BinaryContent:
 			out = genai.Blob{MIMEType: p.MIMEType, Data: p.Data}
 		case llms.ImageURLContent:
-			typ, data, err := DownloadImageData(p.URL)
+			typ, data, err := util.DownloadImageData(p.URL)
 			if err != nil {
 				return nil, err
 			}
