+++ llms/googleai/vertex/vertex.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"cloud.google.com/go/vertexai/genai"
+	"github.com/tmc/langchaingo/internal/util"
 	"github.com/tmc/langchaingo/llms"
-	"github.com/tmc/langchaingo/llms/googleai"
 	"github.com/tmc/langchaingo/schema"
 	"google.golang.org/api/iterator"
 )
@@ -130,7 +130,7 @@ func convertParts(parts []llms.ContentPart) ([]genai.Part, error) {
 		case llms.BinaryContent:
 			out = genai.Blob{MIMEType: p.MIMEType, Data: p.Data}
 		case llms.ImageURLContent:
-			typ, data, err := googleai.DownloadImageData(p.URL)
+			typ, data, err := util.DownloadImageData(p.URL)
 			if err != nil {
 				return nil, err
 			}
